EDIT: dropped & created database/migration exceptions

+++ db/migrate/20240404183232_drop_owner.rb
@@ -1,5 +1,7 @@
 class DropOwner < ActiveRecord::Migration[7.1]
   def change
-    drop_table :owners
+    if table_exists?(:owners)
+      drop_table :owners
+    end
   end
 end
